+++ packages/core/src/constructs.ts
@@ -1453,7 +1453,9 @@ export function or(
       state: DocState<undefined | [number, ParserResult<unknown>]>,
       _defaultValue?,
     ) {
+      let brief: Message | undefined;
       let description: Message | undefined;
+      let footer: Message | undefined;
       let fragments: readonly DocFragment[];
 
       if (state.kind === "unavailable" || state.state == null) {
@@ -1471,7 +1473,9 @@ export function or(
           innerState,
           undefined,
         );
+        brief = docFragments.brief;
         description = docFragments.description;
+        footer = docFragments.footer;
         fragments = docFragments.fragments;
       }
       const entries: DocEntry[] = fragments.filter((f) => f.type === "entry");
@@ -1485,7 +1489,9 @@ export function or(
         }
       }
       return {
+        brief,
         description,
+        footer,
         fragments: [
           ...sections.map<DocFragment>((s) => ({ ...s, type: "section" })),
           { type: "section", entries },
@@ -1978,6 +1984,7 @@ export function longestMatch(
       state: DocState<undefined | [number, ParserResult<unknown>]>,
       _defaultValue?,
     ) {
+      let brief: Message | undefined;
       let description: Message | undefined;
       let footer: Message | undefined;
       let fragments: readonly DocFragment[];
@@ -1993,6 +2000,7 @@ export function longestMatch(
           const docResult = parsers[i].getDocFragments(
             { kind: "available", state: result.next.state },
           );
+          brief = docResult.brief;
           description = docResult.description;
           footer = docResult.footer;
           fragments = docResult.fragments;
@@ -2003,7 +2011,7 @@ export function longestMatch(
         }
       }
 
-      return { description, fragments, footer };
+      return { brief, description, fragments, footer };
     },
   };
 }